Remove unused helper functions, move specific functions to controlflow and add docstring
