+++ mlbench_core/controlflow/pytorch/__init__.py
@@ -1,15 +1,15 @@
 from .checkpoints_evaluation import CheckpointsEvaluationControlFlow
 from .controlflow import (
-    prepare_batch,
     record_train_batch_stats,
     record_validation_stats,
     validation_round,
 )
+from .helpers import prepare_batch
 
 __all__ = [
     "CheckpointsEvaluationControlFlow",
-    "prepare_batch",
     "record_validation_stats",
     "record_train_batch_stats",
     "validation_round",
+    "prepare_batch",
 ]
